pyproject.toml
[project]
name = "brandpilot"
version = "0.1.0"
description = "AI-powered Instagram management platform with multi-agent automation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "black>=26.3.1",
    "fal-client>=0.13.2",
    "httpx>=0.28.1",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "psycopg2-binary>=2.9.10",
    "requests>=2.33.0",
    "schedule>=1.2.2",
    "tavily-python>=0.7.23",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.20",
    "itsdangerous>=2.2.0",
    "cloudinary>=1.40.0",
    "pillow>=12.2.0",
]
